+++ CrazyCanvas/Include/World/Level.h
@@ -3,7 +3,7 @@
 #include "LambdaEngine.h"
 
 #include "LevelModule.h"
-#include "Game/World/LevelObjectCreator.h"
+#include "LevelObjectCreator.h"
 
 #include "Containers/THashTable.h"
 #include "Containers/TArray.h"
@@ -25,6 +25,7 @@ class Level : public LambdaEngine::IClientEntityAccessor
 	{
 		LambdaEngine::TArray<uint64> SaltUIDs;
 		LambdaEngine::TArray<LambdaEngine::Entity> Entities;
+		LambdaEngine::TArray<LambdaEngine::TArray<LambdaEngine::Entity>> ChildEntities;
 	};
 
 public:
@@ -39,7 +40,7 @@ public:
 	// Server & Client Loads Level, Client only loads clientside level objects,
 	// Server then sends packages about server side entities that need to be created in the client, those should use this method
 	// This method should delegate to LevelObjectCreator
-	bool CreateObject(LambdaEngine::ESpecialObjectType specialObjectType, void* pData);
+	bool CreateObject(ESpecialObjectType specialObjectType, void* pData);
 
 	/*
 	*	Spawns a player at a random Spawnpoint, the player is forced to be local
@@ -49,13 +50,13 @@ public:
 		const LambdaEngine::AnimationComponent& animationComponent,
 		const LambdaEngine::CameraDesc* pCameraDesc);
 
-	LambdaEngine::Entity* GetEntities(LambdaEngine::ESpecialObjectType specialObjectType, uint32& countOut);
+	LambdaEngine::Entity* GetEntities(ESpecialObjectType specialObjectType, uint32& countOut);
 
 private:
 	virtual LambdaEngine::Entity GetEntityPlayer(uint64 saltUID) override;
 
 private:
 	LambdaEngine::String m_Name = "";
-	LambdaEngine::THashTable<LambdaEngine::ESpecialObjectType, uint32> m_EntityTypeMap;
+	LambdaEngine::THashTable<ESpecialObjectType, uint32> m_EntityTypeMap;
 	LambdaEngine::TArray<LevelEntitiesOfType> m_Entities;
 };
