+++ CrazyCanvas/Source/States/ServerState.cpp
@@ -88,7 +88,7 @@ bool ServerState::OnClientConnected(const LambdaEngine::ClientConnectedEvent& ev
 		.pClient	= pClient,
 		.Position	= position,
 		.Forward	= glm::normalize(glm::vec3(1.0f, 0.0f, 0.0f)),
-		.Scale		= glm::vec3(0.01f),
+		.Scale		= glm::vec3(1.0f),
 	};
 
 	m_pLevel->CreateObject(ESpecialObjectType::SPECIAL_OBJECT_TYPE_PLAYER, &createPlayerDesc);
