+++ CrazyCanvas/Source/World/LevelManager.cpp
@@ -1,5 +1,5 @@
 #include "World/LevelManager.h"
-#include "Game/World/LevelObjectCreator.h"
+#include "World/LevelObjectCreator.h"
 #include "World/LevelModule.h"
 #include "World/Level.h"
