+++ LambdaEngine/Include/Game/ECS/Systems/Rendering/RenderSystem.h
@@ -29,6 +29,10 @@
 
 #include "Rendering/LightRenderer.h"
 
+#include "Game/ECS/Components/Physics/Transform.h"
+#include "Game/ECS/Components/Rendering/AnimationComponent.h"
+#include "Game/ECS/Components/Rendering/MeshComponent.h"
+
 namespace LambdaEngine
 {
 	class Window;
@@ -239,7 +243,7 @@ namespace LambdaEngine
 		void SetRenderStageSleeping(const String& renderStageName, bool sleeping);
 
 		RenderGraph*	GetRenderGraph()			{ return m_pRenderGraph;	}
-		uint64			GetFrameIndex() const	 	{ return m_FrameIndex; }
+		uint64			GetFrameIndex() const	 	{ return m_FrameIndex;		}
 		uint64			GetModFrameIndex() const	{ return m_ModFrameIndex;	}
 		uint32			GetBufferIndex() const	 	{ return m_BackBufferIndex; }
 
@@ -249,15 +253,12 @@ namespace LambdaEngine
 	private:
 		RenderSystem() = default;
 
-		glm::mat4 CreateEntityTransform(Entity entity);
+		glm::mat4 CreateEntityTransform(Entity entity, const glm::bvec3& rotationalAxes);
+		glm::mat4 CreateEntityTransform(const PositionComponent& positionComp, const RotationComponent& rotationComp, const ScaleComponent& scaleComp, const glm::bvec3& rotationalAxes);
 		
-		void OnEntityAdded(Entity entity);
-		void OnEntityRemoved(Entity entity);
-
+		void OnStaticMeshEntityAdded(Entity entity);
 		void OnAnimatedEntityAdded(Entity entity);
-		void OnAnimatedEntityRemoved(Entity entity);
-
-		void AddEntityInstance(Entity entity, GUID_Lambda meshGUID, GUID_Lambda materialGUID, const glm::mat4& transform, bool animated);
+		void OnPlayerEntityAdded(Entity entity);
 
 		void OnDirectionalEntityAdded(Entity entity);
 		void OnDirectionalEntityRemoved(Entity entity);
@@ -265,10 +266,13 @@ namespace LambdaEngine
 		void OnPointLightEntityAdded(Entity entity);
 		void OnPointLightEntityRemoved(Entity entity);
 
-		void RemoveEntityInstance(Entity entity);
+		void AddRenderableEntity(Entity entity, GUID_Lambda meshGUID, GUID_Lambda materialGUID, const glm::mat4& transform, bool animated);
+		void RemoveRenderableEntity(Entity entity);
+
 		void UpdateDirectionalLight(const glm::vec4& colorIntensity, const glm::vec3& position, const glm::quat& direction, float frustumWidth, float frustumHeight, float zNear, float zFar);
 		void UpdatePointLight(Entity entity, const glm::vec3& position, const glm::vec4& colorIntensity, float nearPlane, float farPlane);
-		void UpdateTransform(Entity entity, const glm::mat4& transform);
+		void UpdateAnimation(Entity entity, MeshComponent& meshComp, AnimationComponent& animationComp);
+		void UpdateTransform(Entity entity, const PositionComponent& positionComp, const RotationComponent& rotationComp, const ScaleComponent& scaleComp, const glm::bvec3& rotationalAxes);
 		void UpdateCamera(const glm::vec3& position, const glm::quat& rotation, const CameraComponent& camComp, const ViewProjectionMatricesComponent& viewProjComp);
 
 		void CleanBuffers();
@@ -291,11 +295,12 @@ namespace LambdaEngine
 		void UpdateRenderGraph();
 
 	private:
+		IDVector m_StaticMeshEntities;
+		IDVector m_AnimatedEntities;
+		IDVector m_PlayerEntities;
 		IDVector m_DirectionalLightEntities;
 		IDVector m_PointLightEntities;
-		IDVector m_RenderableEntities;
 		IDVector m_CameraEntities;
-		IDVector m_AnimatedEntities;
 
 		TSharedRef<SwapChain>	m_SwapChain			= nullptr;
 		Texture**				m_ppBackBuffers		= nullptr;
