+++ LambdaEngine/Source/Resources/ResourceLoader.cpp
@@ -1409,7 +1409,9 @@ namespace LambdaEngine
 					LoadSkeleton(pMesh, pMeshAI);
 					if (pMesh->pSkeleton)
 					{
-						glm::mat4 meshTransform		= AssimpToGLMMat4(pNode->mTransformation);
+						//Assume Mixamo has replaced our rotation
+						glm::mat4 meshTransform		= glm::rotate(glm::mat4(1.0f), glm::pi<float>(), glm::vec3(0.0f, 1.0f, 0.0f));
+						meshTransform				= meshTransform * AssimpToGLMMat4(pNode->mTransformation);
 					    glm::mat4 globalTransform	= AssimpToGLMMat4(pScene->mRootNode->mTransformation);
 					    pMesh->pSkeleton->InverseGlobalTransform = glm::inverse(globalTransform) * meshTransform;
